Added inital workings on using the cake build system on mineserver.
The cake build system is available from http://cake-build.sourceforge.net/
It is python based so requires python installed on the system.

For the curious:

$ git clone git://cake-build.git.sourceforge.net/gitroot/cake-build/cake-build
$ cd cake-build
$ ./setup.py install

I have not yet tested under GNU/Linux just yet, on Microsoft Windows you will need your python's script folder in the path in order to run cake.

Once it is fininshed and if it is accepted, I will write up a page about how to set it up and use it.